更新blueprint
