+++ api/__init__.py
@@ -1,0 +1,5 @@
+from api.attraction_api import attraction_api
+from api.signup_api import signup_api
+from api.categories_api import categories_api
+from api.login_api import login_api
+from api.booking_api import booking_api
