+++ static/styles/booking.css
@@ -2,6 +2,12 @@
     margin:0px;
     padding:0px;
 }
+body{
+    font-family: "Noto Sans TC", "微軟正黑體", sans-serif;
+}
+p{
+    font-family: "Noto Sans TC", "微軟正黑體", sans-serif;
+}
 .booking-content{
     padding-top: 100px;
     width:1200px;
@@ -38,7 +44,7 @@
 }
 .right-product-title{
     color: #448899;
-    font-family: 'Noto Sans TC';
+    font-family: "Noto Sans TC", "微軟正黑體", sans-serif;
     font-style: normal;
     font-weight:700;
     margin-bottom:20px;
@@ -150,7 +156,7 @@ input{
 .total-price{
     text-align: center;
     margin:auto;
-    font-family: 'Noto Sans TC';
+    font-family: "Noto Sans TC", "微軟正黑體", sans-serif;
     font-style: normal;
     font-weight: 700;
     font-size: 16px;
@@ -162,7 +168,7 @@ input{
     margin:auto;
     background-color: #448899;
     color:white;
-    font-family: 'Noto Sans TC';
+    font-family: "Noto Sans TC", "微軟正黑體", sans-serif;
     font-style: normal;
     font-weight: 700;
     font-size: 16px;
